Removed unused handlers. Moved auth helper functions to handlers sub-folder
