+++ src/app/components/friendsComponents/AddFriendPopover.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { emailValidation } from "@/app/lib/auth/emailValidation";
+import { emailValidation } from "@/app/lib/handlers/auth_handlers/emailValidation";
 import { UserType } from "@/app/lib/handlers/getUser";
 import {
   Button,
