+++ src/app/lib/handlers/auth_handlers/emailValidation.ts
