+++ src/app/lib/handlers/auth_handlers/passwordValidation.ts
