--- src/app/lib/handlers/loginHandler.ts
@@ -1,11 +1,0 @@
-export type loginBody = {
-  email: string;
-  password: string;
-};
-
-export type resType = {
-  msg: "success" | "error";
-  data?: {
-    errorMessage?: string;
-  };
-};
