--- src/app/lib/handlers/logoutHandler.ts
@@ -1,7 +1,0 @@
-// import { signOut } from "next-auth/react";
-
-import { signOut } from "@/auth";
-
-export const logoutHandler = async () => {
-  await signOut({ redirectTo: "/apps", redirect: true });
-};
